ML_django/requirements.txt
alembic==1.3.1
aniso8601==8.0.0
application==1.0
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
django-storages==1.8
djangorestframework==3.10.3
Flask==1.1.1
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.1
google==2.0.2
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-oauth==1.0.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==2.13.6
marshmallow-sqlalchemy==0.15.0
MyApplication==0.1.0
mysqlclient==1.3.13
Pillow==6.2.1
protobuf==3.11.0
psycopg2-binary==2.7.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.2.13
sqlparse==0.3.0
urllib3==1.25.7
Werkzeug==0.16.0
